refactor: move upload exceptions to plugin/upload and add FileUploadError
Moves UploadError and BackendUnavailableError from OTCamera/exceptions.py
into OTCamera/plugin/upload/exceptions.py and introduces FileUploadError
(inheriting UploadError) for file-level upload failures. Both FtpUpload
and S3Upload now raise FileUploadError. All three exceptions are
re-exported from OTCamera/plugin/upload/__init__.py.
