+++ OTCamera/plugin/upload/__init__.py
@@ -1,1 +1,9 @@
 """Upload plugin implementations."""
+
+from OTCamera.plugin.upload.exceptions import (
+    BackendUnavailableError,
+    FileUploadError,
+    UploadError,
+)
+
+__all__ = ["UploadError", "FileUploadError", "BackendUnavailableError"]
