+++ OTCamera/plugin/upload/ftp_upload.py
@@ -5,7 +5,7 @@ from ftplib import FTP_TLS
 from pathlib import Path
 
 from OTCamera.domain.upload import Upload
-from OTCamera.exceptions import UploadError
+from OTCamera.plugin.upload.exceptions import FileUploadError
 
 logger = logging.getLogger(__name__)
 
@@ -44,7 +44,7 @@ class FtpUpload(Upload):
             with open(source, "rb") as file_handle:
                 client.storbinary(f"STOR {destination.name}", file_handle)
         except Exception as exc:
-            raise UploadError(f"Upload failed for {source.name}: {exc}") from exc
+            raise FileUploadError(f"Upload failed for {source.name}: {exc}") from exc
         finally:
             client.close()
